simplifying main func, creating helper classes
