+++ internal/common/model.go
@@ -1,0 +1,8 @@
+package common
+
+type S3Connection struct {
+	BucketName             string
+	DefaultPrefixToPrepend string
+	ProfileName            string
+	DefaultStorageClass    string
+}
